[P] Implement the first iteration of the Timing tab.
The types are:

type Map = {
    timing: TimingSegment[]; // sorted by start time
};
type TimingSegment = {
    startTime: number;
    bpm: number;
};

The canvas should be replaced with a list of timing segments. You should be able to click a timing segment and edit it in a right sidebar. You should be able to add a timing segment. The selected song should have its player at the top, and adding a timing segment should add it at the current time in the song.